pyproject.toml
[project]
name = "vision-interpret"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "transformers>=5.2.0",
]
